+++ src/components/IconButton.tsx
@@ -1,5 +1,5 @@
 import { Pressable, StyleSheet } from "react-native";
-import { Ionicons } from "@expo/vector-icons";
+import Ionicons from "@expo/vector-icons/Ionicons";
 import { theme } from "../theme/theme";
 
 type Props = {